COOP: add reporting to redirects
This CL allows reporting browsing context group switches triggered by
redirects and updates the reports sent in this case to the latest
version of the spec PR (https://github.com/whatwg/html/pull/5518). Since
the status of COOP was becoming hard to track, I moved it to its own
class for better encapsulation.

Bug: 1059303
Change-Id: Ifafb23073301bd05cd9ce83fdb0b748c28e8a51f